requirements.txt
absl-py==2.3.1
Box2D==2.3.10
cloudpickle==3.1.2
Farama-Notifications==0.0.4
filelock==3.20.0
fsspec==2025.12.0
grpcio==1.76.0
gymnasium==1.2.2
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.10.0.84
packaging==25.0
pillow==12.0.0
protobuf==6.33.2
pygame==2.6.1
PyYAML==6.0.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.9.1
triton==3.5.1
typing_extensions==4.15.0
Werkzeug==3.1.4
